+++ CHANGELOG.md
@@ -5,9 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
 
-## 1.0.0-beta - 2024-04-15
+### Changed
+
+- Refactored IMAP auth config API
+
+  The IMAP auth config option is now explicit, in order to improve error messages:
+
+  ```toml
+  # before
+  right.backend.password.cmd = "pass show example"
+
+  # after
+  right.backend.auth.type = "password"
+  right.backend.auth.cmd = "pass show example"
+  ```
+
+## [1.0.0-beta] - 2024-04-15
 
 ### Added
 
@@ -24,8 +39,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `envelope.filter` to `envelope.filters` in order to match lib types.
 - Renamed `check` command to `doctor`.
 
-## 0.1.0 - 2024-04-10
+## [0.1.0] - 2024-04-10
 
 ### Added
 
-- Initiated the project from [Himalaya CLI](https://github.com/soywod/himalaya).
+- Initiated the project from [Himalaya CLI](https://github.com/pimalaya/himalaya).
+
+[Unreleased]: https://github.com/pimalaya/neverest/compare/v1.0.0-beta...HEAD
+[1.0.0-beta]: https://github.com/pimalaya/neverest/compare/v0.1.0...v1.0.0-beta
+[0.1.0]: https://github.com/pimalaya/neverest/compare/root...v0.1.0